feat: Verify annotations in key, email, url verifiers
For PublicKeyVerifier, CertSubjectEmailVerifier, CertSubjectUrlVerifier,
add a new `annotations` parameter, and logic so they check that their
layer contains the passed annotations.

This way, we can for example create a PublicKeyVerifier with a key and
annotations. When filtering with cosign::filter_signature_layers(), we
will filter for those signatures signed with that key and containing
those annotations.

This is in contrast with the already existing AnnotationsVerifier, that
is used to filter for those signatures containing the
annotations (regardless of the identity), and it's used to check that
all signatures contain the passed annotations.

Consume changes to verifiers.

Co-authored-by: Rafael Fernández López <rfernandezlopez@suse.com>
Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>
Signed-off-by: Rafael Fernández López <rfernandezlopez@suse.com>