fix!: use all Fulcio's certificates
Fulcio has recently introduced a new root CA to issue certificates.
Prior to this commit, the code was making use only of the initial
root CA.
Because of that, recent keyless signatures were considered not valid.

Now the code is taking into account of the root CAs that Fulcio used
over the time.

This fixes https://github.com/sigstore/sigstore-rs/issues/32

Signed-off-by: Flavio Castelli <fcastelli@suse.com>