refactor!: Return SigstoreVerifyConstraintsError in verify_constraints()
`verify_constraints()` now returns Ok(), or an error with a vector of
references to unsatisfied constraints (instead of a vector of
constraints without references). This simplifies ownership of the unsatisfied
constraints.

Refactor to accept an iter() of constraints instead of the
VerificationConstraintsVec. This removes the need of deriving Clone for
VerificationConstraint.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>
Co-Authored-by: Flavio Castelli <fcastelli@suse.com>
Signed-off-by: Flavio Castelli <fcastelli@suse.com>