Update link to verification example
This commit updates the link in examples/cosign/verify/README.md to
point to main.rs.

The motivation for this is that the current link will return a 404 "Not
Found" because the path is missing 'cosign' in it. Adding 'cosign' will
resolve the not found issue, but in that case it is just linking back
to the same README.md.

Hopefully, linking to main.rs will make things clear about which code
is being run in the cargo command that follows after the link.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>