Set tag to COSIGN_PRIVATE_KEY_PEM_LABEL for ec.rs
This commit suggests changing the tag used in
EcdsaKeys::private_key_to_encrypted_pem from
`SIGSTORE_PRIVATE_KEY_PEM_LABEL` to `COSIGN_PRIVATE_KEY_PEM_LABEL`.

The motivation for this is that using such a key with Go cosign will
currently result in an `unsupported pem type` error.
It seems like Go cosign is is using
[CosignPrivateKeyPemType](https://github.com/sigstore/cosign/blob/6b309df06f60ea5f58db22e9890713138c823d27/pkg/cosign/keys.go#L41) when [checking](https://github.com/sigstore/cosign/blob/6b309df06f60ea5f58db22e9890713138c823d27/pkg/cosign/keys.go#L208) the tag of the pem.

I'm not sure about the differences between
SIGSTORE_PRIVATE_KEY_PEM_LABEL and COSIGN_PRIVATE_KEY_PEM_LABEL and I
might be missing something here but if the change in this PR is a valid
then this should probably be done for rsa and ed25519aswell.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>