fix: make LogEntry Body an enum
This commit updates the LogEntry body field type to be an enum instead
of a struct. This enum will contain a variant for the different types of
rekords that can be stored in Rekor.

The motivation for this change is that currently only Hashedrekord::Spec
are supported as that is the type spec field in the current Body struct.
For further details please see the linked issue.

Closes: https://github.com/sigstore/sigstore-rs/issues/243
Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>