chore(deps): massive update of crypto libraries
Update to the latest version of a bunch of crypto libraries. These
libraries, especially the ones that belong to RustCrypto, have to be
updated at the same time.

There have been a lot of API changes and deprecation warnigns, which needed
quite some adaptation of our codebase.

This fixes the following issues: #267, #266, #265, #248, #236, #234, #233, #230

Signed-off-by: Flavio Castelli <fcastelli@suse.com>