fix: fetch Sigstore's TUF repository when default flags are disabled
Fetching Sigstore's TUF repository is done using the `aws/tough` crate.
This crate is currently using an older version of `reqwest`, which in turn
uses an older version of `hyper`.

When building the `sigstore` crate with the default features turned off and
a limited set of "verification-related" features enabled (like
`sigstore-trust-root` and `cosign-rustls-tls`), fetching the TUF
repository causes a runtime panic. The panic happens because the old
`reqwest` crate is built without TLS support.
This was not a problem before, since both sigstore and `tough` used the
same version of `reqwest`, hence enabling `cosign-rustls-tls` led to
TLS support being enabled also for tough.

This commit introduces two new feature flags:
`sigstore-trust-root-native-tls` and `sigstore-trust-root-rustl-tls`
which enable TLS support also for the old version of reqwest being
currently used by tough.

Worth of interest, building with the default flags doesn't currently
expose this issue. That happens because of a happy coincidence.
The "feature chain reaction" is the following one:
- The "enable default features" cause the `full-native-tls` feature to be enabled
- The `full-native-tls` feature enables `fulcio-native-tls` feature
- The `fulcio-native-tls` features enables the `oauth-native-tls` feature
- The `oauth-native-tls` features causes the `openidconnect` crate to be required,
  moreover it enables TLS support of `openidconnect` by enabling the
  `openidconnect/native-tls` feature.

Currently the `openidconnect` crate depends on the same old version of `reqwest` used by
aws/tough. Hence, enabling TLS support for `openidconnect` leads to TLS support being
enabled also by tough.
This coincidental fix is going to disappear as soon as `openidconnect`
moves to the latest version of `reqwest`.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>