ci: Test all feature combinations
This uses cargo-hack to test all possible feature combinations.
Since at present this is ~800 combinations, cargo check has been
split to operation in parallel.

Single features are tested in isolation (with a few exceptions)
All features combinations are tested with `native-tls`
All features combinations are tested with `rustls-tls`

This generates a lot of compiler warnings for dead-code since some
modules expose methods that are only accessed when other features are
enabled.

Given the combination of features is so large, cargo test is run twice
only on `full,native-tls` and `full,rustls-tls`.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>