chore(deps): reorganize wasm dependencies
The getrandom crate is used only by the wasm feature, through the
`ring` dependency. It was explicitly mentioned by our Cargo.toml to
enable the `js` feature, required by the `wasm32-unknown-unknown` target.

Starting from this commit, we stop mentioning the
`getrandom` dependency and instead we make the `ring` dependency
explicit. The `js` feature of the getrandom crate is now set by using
a dedicated feature of the ring crate itself.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>